added authorization for update and delete campground

+++ routes/campgrounds.js
@@ -5,6 +5,7 @@ const ExpressError = require('../helpers/ExpressError')
 const Campground = require('../models/campground')
 const { campgroundSchema } = require('../schemas')
 const { isLoggedIn } = require('../middleware')
+const campground = require('../models/campground')
 
 const validateCampground = (req, res, next) => {
     const { error } = campgroundSchema.validate(req.body)
@@ -39,7 +40,6 @@ router.get('/', catchAsync(async (req, res) => {
 //show single campground
 router.get('/:id', catchAsync(async (req, res) => {
     const campground = await Campground.findById(req.params.id).populate('reviews').populate('author')
-    console.log(campground)
     if (!campground) {
         req.flash('error', 'Cannot find that campground!')
         return res.redirect('/campgrounds')
@@ -49,10 +49,15 @@ router.get('/:id', catchAsync(async (req, res) => {
 
 //show edit campground form
 router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id)
+    const { id } = req.params
+    const campground = await Campground.findById(id)
     if (!campground) {
         req.flash('error', 'Cannot find that campground!')
         return res.redirect('/campgrounds')
+    }
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('error', 'You do not have permission to do that')
+        return res.redirect(`/campgrounds/${id}`)
     }
     res.render('campgrounds/edit', { campground })
 }))
@@ -60,7 +65,12 @@ router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
 //update campground
 router.put('/:id', isLoggedIn, catchAsync(async (req, res) => {
     const { id } = await req.params
-    const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground })
+    const campground = await Campground.findById(id)
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('error', 'You do not have permission to do that')
+        return res.redirect(`/campgrounds/${id}`)
+    }
+    const camp = await Campground.findByIdAndUpdate(id, { ...req.body.campground })
     req.flash('success', 'Successfully updated this campground!')
     res.redirect(`/campgrounds/${campground._id}`)
 }))
@@ -68,6 +78,11 @@ router.put('/:id', isLoggedIn, catchAsync(async (req, res) => {
 //delete campground
 router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id)
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('error', 'You do not have permission to do that')
+        return res.redirect(`/campgrounds/${id}`)
+    }
     await Campground.findByIdAndDelete(id);
     req.flash('success', 'Successfully deleted campground!')
     res.redirect('/campgrounds');
